% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_level_methods.R
\name{format.auto_segment_levels}
\alias{format.auto_segment_levels}
\title{Format an `auto_segment_levels` Result}
\usage{
\method{format}{auto_segment_levels}(x, ...)
}
\arguments{
\item{x}{An object of class `"auto_segment_levels"`.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A single character string.
}
\description{
One-line summary suitable for `cat()` contexts.
}
