% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_methods.R
\name{format.moneca_bipartite}
\alias{format.moneca_bipartite}
\title{Format a `moneca_bipartite` object}
\usage{
\method{format}{moneca_bipartite}(x, ...)
}
\arguments{
\item{x}{An object of class `"moneca_bipartite"`.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A single character string describing the bipartite fit.
}
\description{
One-line summary of a bipartite fit suitable for `cat()` contexts.
}
