% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_methods.R
\name{print.summary.moneca_bipartite}
\alias{print.summary.moneca_bipartite}
\title{Print a `summary.moneca_bipartite` object}
\usage{
\method{print}{summary.moneca_bipartite}(x, ...)
}
\arguments{
\item{x}{An object of class `"summary.moneca_bipartite"`.}

\item{...}{Further arguments, currently ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a `summary.moneca_bipartite` object
}
